@@ -12,9 +12,19 @@ public class RobloxAfker {
 
     @SneakyThrows
     private RobloxAfker(String[] args) {
+        boolean randomWalking = false;
+
+        if (args.length == 0) {
+            System.out.println("Please enter a argument.");
+            System.exit(0);
+        } else if (args[0].equalsIgnoreCase("randomWalking")) {
+            System.out.println("Starting randomWalking...");
+            randomWalking = true;
+        }
+
         new KeyEvent(this);
 
         Timer timer = new Timer("okokok");
-        timer.scheduleAtFixedRate(new KeyRunnable(this), 5000L, 1000L);
+        timer.scheduleAtFixedRate(new KeyRunnable(this, randomWalking), 5000L, 1000L);
     }
 }
